Adds custom bash_it script to update bash history
By default each bash sessions history diverges when the shell is
created and the history is only saved when the shell exits.

This custom plugin causes the bash history to be updated whenever the
command prompt is rendered This means all bash windows will use a
constantly updated shared history.

Signed-off-by: Thinh Nguyen <thinh.nguyen@emc.com>